% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ax.R
\name{ax}
\alias{ax}
\title{Draw Pretty Axes with Tickmarks}
\usage{
ax(side = 1, tick = 100, ticksize = -0.01, labs = NULL, ...)
}
\arguments{
\item{side}{An integer specifying which side of the plot the axis is to be drawn on.
The value can be 1 (bottom), 2 (left), 3 (top), or 4 (right). Default is 1.}

\item{tick}{A numeric value specifying the interval between tick marks. Default is 100.}

\item{ticksize}{A numeric value specifying the size of the tick marks. Default is -0.01 (negative values indicate a position below the axis).}

\item{labs}{A character vector specifying the labels for the tick marks. If \code{NULL}, labels are automatically generated. Default is \code{NULL}.}

\item{...}{Additional arguments to be passed to the \code{rug} function.}
}
\value{
NULL. This function is used for its side effect of creating an axis on a plot.
}
\description{
This function draws axes with minor-interval tick marks and labels, allowing for customization of tick mark intervals and sizes.
}
\examples{
# Example usage
plot(1:10, 1:10, type = "n")
ax(side = 1, tick = 1, col=2)
ax(side = 2, tick = 1, ticksize = -0.05, labs = letters[1:10])
}
\seealso{
\code{\link{rug}}
}
