% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/battleship.MCd.R
\name{battleship.MCd}
\alias{battleship.MCd}
\title{Plot Density Models as "Battleship" Plots}
\usage{
battleship.MCd(
  x,
  os = 0,
  add = FALSE,
  col = rgb(0, 0, 0.8, 0.5),
  lcol = rgb(0, 0, 0.8, 0.8),
  lwd = 1,
  lty = 1,
  scalefactor = 1,
  ylab = "Density",
  xlab = "Automatic",
  grid = TRUE,
  ...
)
}
\arguments{
\item{x}{A matrix or data frame, where the first column represents the x-axis values (e.g., years)
and the remaining columns represent density values from Monte Carlo simulations.}

\item{os}{Numeric. An offset to adjust the baseline of the plot. Default is 0.}

\item{add}{Logical. If \code{TRUE}, the graph is added to an existing plot. Default is \code{FALSE}.}

\item{col}{Color of the filled area representing the density envelope. Default is semi-transparent blue (\code{rgb(0, 0, 0.8, 0.5)}).}

\item{lcol}{Line color for the edges of the density envelope. Default is blue (\code{rgb(0, 0, 0.8, 0.8)}).}

\item{lwd}{Line width for the edges of the density envelope. Default is 1.}

\item{lty}{Line type for the edges of the density envelope. Default is 1 (solid line).}

\item{scalefactor}{Numeric. A multiplier applied to scale the density values. Default is 1.}

\item{ylab}{Character. Label for the y-axis. Default is \code{"Density"}.}

\item{xlab}{Character. Label for the x-axis. If \code{"Automatic"}, labels are determined based on the x-axis values (e.g., \code{"Cal. BC/AD"}). Default is \code{"Automatic"}.}

\item{grid}{Logical. If \code{TRUE}, a grid is added to the plot background. Default is \code{TRUE}.}

\item{...}{Additional graphical parameters passed to the underlying \code{plot} function.}
}
\description{
This function visualizes density models in a style reminiscent of "battleship" seriation graphs.
It plots the mean densities along with their standard deviations, mirrored along the y-axis for symmetry.
}
\details{
The function computes the mean (\code{M}) and standard deviation (\code{Sd}) of the density values for each x-axis value
and visualizes them as a polygon envelope centered around the baseline (\code{os}). The envelope is symmetric around the baseline
and bounded by \verb{M ± Sd}. Optionally, it adds the plot to an existing graph using the \code{add} parameter.

If \code{xlab} is set to \code{"Automatic"}, the function attempts to infer the appropriate x-axis label based on the range of x-axis values.
It uses \code{"Cal. AD"} for entirely positive ranges, \code{"Cal. BC"} for entirely negative ranges, and \code{"Cal. BC/AD"} for mixed ranges.
}
\examples{
dates <- data.frame(BP = c(4840, 4885, 4739, 4826, 4610), sd = c(45, 50, 27, 24, 31))
denmod <- MCdensity(dates)
# Plot the battleship graph
battleship.MCd(denmod, col = rgb(0.2, 0.5, 0.9, 0.4), lcol = "darkblue")

}
\seealso{
\link{battleships}
}
