% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/battleships.R
\name{battleships}
\alias{battleships}
\title{Plot Multiple "Battleship" Density Curves}
\usage{
battleships(
  L,
  cols = rep(c(1:9), 10),
  space = 1.5,
  ylas = 1,
  xlab = "Automatic",
  ylab = "",
  ...
)
}
\arguments{
\item{L}{An object of class \code{MCd} or a named list of \code{MCd} objects to plot.}

\item{cols}{A vector of colors for the "battleship" curves. Defaults to a repeating sequence of integers (\code{rep(c(1:9),10)}).}

\item{space}{Numeric. Vertical spacing between the "battleship" curves. Default is \code{1.5}.}

\item{ylas}{Numeric. Orientation of the y-axis labels. See \code{las} in \link{par}. Default is \code{1} (horizontal labels).}

\item{xlab}{Character. Label for the x-axis. If \code{"Automatic"}, the function infers the label based on the x-axis values. Default is \code{"Automatic"}.}

\item{ylab}{Character. Label for the y-axis. Default is an empty string (\code{""}).}

\item{...}{Additional graphical parameters passed to the underlying \code{plot} function.}
}
\description{
This function visualizes one or multiple Monte Carlo density (MCd) models as "battleship" seriation graphs,
allowing comparisons across different models.
}
\details{
The \code{battleships} function plots multiple density models as stacked "battleship" seriation-style graphs. Each "battleship" represents
the mean density and its variation (as standard deviations) for a given \code{MCd} model. If \code{L} is a single \code{MCd} object, it is
treated as a single entry in a list.

The function automatically determines the x-axis range based on the age ranges in the input data and scales the y-axis
to accommodate all models. The \code{cols} parameter allows customizing the colors for each density model.

When \code{xlab} is \code{"Automatic"}, the function labels the x-axis based on the range of values (e.g., \code{"Cal. AD"}, \code{"Cal. BC"}, or \code{"Cal. BC/AD"}).
If \code{L} is a named list, the names are used as labels for the y-axis.
}
\examples{
dens1 <- MCdensity(data.frame(
		BP = c(4840, 4885, 4739, 4826, 4610), 
		sd = c(45, 50, 27, 24, 31)))
dens2 <- MCdensity(data.frame(
		BP = c(4890, 4950, 4820, 4826, 4610), 
		sd = c(45, 23, 27, 25, 35)))
battleships(list("Region 1" = dens1, "Region 2" = dens2), cols = c("blue", "red"))

}
\seealso{
\link{battleship.MCd}, \link{plot}, \link{plot.MCd}
}
