% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrap_density.R
\name{bootstrap_density}
\alias{bootstrap_density}
\title{Bootstrap a confidence interval for a set of Calendar Dates}
\usage{
bootstrap_density(DATA = CLIP()[, 1], N = 100, bw = 30)
}
\arguments{
\item{DATA}{A vector of calendar dates. BY default the function takes input from the clipboard.}

\item{N}{Number of bootstrap samples.}

\item{bw}{Bandwidth parameter for density calculation.}
}
\value{
A matrix of type 'MCd' containing the \code{MCdensity}-like object.
}
\description{
This function generates a \code{MCdensity}-like object from a vector of calendar dates without uncertainty using bootstrapping (sampling with replacement).
}
\details{
This function is intended for historic records that have no uncertainty associated with them.
An alternative approach is to use the \code{MCdensity} function with a known uncertainty (e.g. 0.5 years) associated with each date.
}
\examples{
# Simulate 50 calendar dates centered on AD 1066 with a sigma of 50
dates<-rnorm(50, 1066, 50)
# calculate and plot bootstrap density model
denmod<-bootstrap_density(dates)
plot(denmod)

#compare to MCdensity with six months of uncertainty and no bootstrapping the data
denmod2<-MCdensity(rowunif(dates-0.5, dates+0.5))
plot(denmod2, add=T, col=2)

}
\references{
McLaughlin, R., Hannah, E., and Coyle-McClung, L. 2018. Frequency analyses of historical and archaeological datasets reveal the same pattern of declining sociocultural activity in 9th to 10th Century CE Ireland. Cliodynamics 9: 1–24. https://doi.org/10.21237/C7clio9136654
}
\seealso{
\code{\link{MCdensity}} \code{\link{mixdensity}} \code{\link{phasedensity}} \code{\link{density}} \code{\link{plot.MCd}}
}
\author{
T. Rowan McLaughlin
}
