% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{calcal}
\alias{calcal}
\title{calcal: a dummy calibration curve for calendar ages from 50,000 BC to 2100 AD.}
\format{
A data.frame with the following fields:
\describe{
\item{\code{calBP}}{Calendar age in years before AD 1950}
\item{\code{BCAD}}{Calendar age in years BC/AD}
\item{\code{Error}}{Age error (0)}
}
}
\usage{
calcal
}
\description{
calcal: a dummy calibration curve for calendar ages from 50,000 BC to 2100 AD.
}
\keyword{datasets}
