% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calen.R
\name{calen}
\alias{calen}
\title{Generate a Normal Distribution of Calendar Dates}
\usage{
calen(date, sigma, sigmas = 2, res = if (sigma < 5) 0.1 else 1)
}
\arguments{
\item{date}{The calendar date forming the mean of the desired}

\item{sigma}{The desired standard deviation}

\item{sigmas}{Number of standard deviations to include in the normal distribution. Default is 2.}

\item{res}{Resolution of the generated calendar dates. Default is 0.1 for sigma less than 5, otherwise 1.}
}
\value{
A matrix with two columns: calendar dates and corresponding probabilities.
}
\description{
This function generates a normal distribution whose mean is the supplied calendar date, with output in a two-column matrix, i.e. the same format used by \code{rowcal}.
}
\details{
If the sigma value is zero, a small uncertainty (0.5) is added to avoid errors.
The function generates a normal distribution  of probability masses around the input date.
The number of sigmas determines the range of calendar dates.
The resolution of the generated dates is adjusted based on the sigma value.
}
\examples{
calen(1000, 2)
calen(200, 3, sigmas = 3, res = 0.5)

}
