% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calsam.R
\name{calsam}
\alias{calsam}
\title{Calendar age samples}
\usage{
calsam(date, sigma, N = 1, ...)
}
\arguments{
\item{date}{The radiocarbon determination.}

\item{sigma}{The sigma value corresponding to the radiocarbon determination.}

\item{N}{Number of calendar dates to sample. Default is 1.}

\item{...}{Additional arguments to be passed to the \code{calen} function.}
}
\value{
A numeric vector containing the sampled calendar date(s).
}
\description{
This function calculates one or more calendar age samples, based on the \code{calen} function.
}
\details{
This function generates a normal distribution of calendar dates using the given mean date and sigma value.
It then samples one or more calendar dates based on the generated distribution.
}
\examples{
calsam(100, 2)
calsam(200, 3, N = 5)

}
