% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CLIP.R
\name{CLIP}
\alias{CLIP}
\title{Input Data from Clipboard Independent of Platform}
\usage{
CLIP()
}
\value{
A data frame containing the data from the clipboard.
}
\description{
This function reads data from the clipboard, independent of the operating system platform.
}
\details{
This function checks the operating system platform and reads data from the clipboard accordingly.
On macOS, it uses the \code{pbpaste} command to read data from the clipboard.
On Windows and Linux, it reads data from the clipboard using the \code{read.delim} function with 'clipboard' as the filename.
}
\examples{
CLIP()

}
