% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fill.R
\name{fill}
\alias{fill}
\title{Fill Polygon with Color}
\usage{
fill(j, col = "grey", border = NA, ...)
}
\arguments{
\item{j}{A matrix containing the coordinates of the polygon vertices.}

\item{col}{The fill color for the polygon (default is "grey").}

\item{border}{The border color for the polygon (default is NA, meaning no border).}

\item{...}{Additional graphical parameters passed to the \code{polygon} function.}
}
\description{
This function fills a polygon defined by a set of coordinates with a specified color.
}
\details{
This function fills a polygon defined by a set of coordinates with a specified color and optional border. It is designed for filling the area under a probabilty density function defined by a two-column matrix, such as the output of \code{rowcal}.
}
\examples{
# Plot a the calibration of 5310±45 BP
test_date<-rowcal(5310, 45)
plot.default(test_date, col=NA)
fill(test_date, col='#CCCCFF', border='#0000FF')

}
\seealso{
\code{\link{plot.MCd}}
}
