% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findmixmedian.R
\name{findmixmedian}
\alias{findmixmedian}
\title{Find the median date for each row of a table of dates with different calibration curves.}
\usage{
findmixmedian(L, ...)
}
\arguments{
\item{L}{A three-column list containing radiocarbon determinations, calendar dates, and/or sapwood dates. Columns are: date, error, and cc (calibration curve name).}

\item{...}{Additional arguments to be passed to the \code{rowcalmedian} function.}
}
\value{
A vector containing the weighted means of samples.
}
\description{
This function returns a list of weighted means of samples, given a three-column input (BP, SD, calibration curve).
}
\details{
This function processes each row of the input list and calculates the weighted mean of samples based on the provided radiocarbon determinations, calendar dates, and sapwood dates.
For calendar dates (cc = 'cal'), the function samples from the calendar dates using the \code{calsam} function.
For radiocarbon determinations, the function calculates the median using the \code{rowcalmedian} function.
}
\examples{
data.frame(date = c(-3800, 4990, 5300), error = c(10, 30, 50), cc = c('cal', 'intcal', 'marine'))
findmixmedian(L)

}
