% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findmode.R
\name{findmode}
\alias{findmode}
\title{Find the mode for Each Row in a list of calibrated dates}
\usage{
findmode(L)
}
\arguments{
\item{L}{A two-column list (radiocarbon determination BP, standard deviation) containing uncalibrated radiocarbon dates.}

\item{...}{Additional arguments to be passed to the \code{rowcalmode} function.}
}
\value{
A vector containing the modal date for each element in the input.
}
\description{
This function generates a best guess for a table of uncalibrated radiocarbon dates using the modal value.
}
\examples{
L<-rowcal(date = c(-3800, 4990, 5300), sigma = c(10, 30, 50),
          cc = c('calcal', 'intcal', 'marine'))
findmode(L)

}
\seealso{
\code{\link{rowcal}} \code{\link{findmedian}} \code{\link{findwm}}
}
\author{
T. Rowan McLaughlin
}
