% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findwm.R
\name{findwm}
\alias{findwm}
\title{Find the Weighted Mean for each element in a list produced by \code{rowcal}}
\usage{
findwm(L, ...)
}
\arguments{
\item{L}{A two-column list (radiocarbon determination BP, standard deviation) containing uncalibrated radiocarbon dates.}

\item{...}{Additional arguments to be passed to the \code{rowcalwm} function.}
}
\value{
A vector containing the best guess weighted mean date for each uncalibrated radiocarbon date.
}
\description{
This function generates a best guess weighted mean for a list of date estimates defined by a probablity distribution.
}
\details{
This function processes each element of the input list and generates the best guess date using weighted mean.
}
\examples{
L<-rowcal(date = c(-3800, 4990, 5300), sigma = c(10, 30, 50),
          cc = c('calcal', 'intcal', 'marine'))
findmm(L)

}
\references{
McLaughlin, T.R. 2019. On applications of space-time modelling with open-source 14C age calibration. Journal of Archaeological Method and Theory 26, 479–501. https://doi.org/10.1007/s10816-018-9381-3
}
\seealso{
\code{\link{rowcal}} \code{\link{findmedian}} \code{\link{findmode}} \code{\link{rowcalwm}}
}
\author{
T. Rowan McLaughlin
}
