% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggrsignif.R
\name{ggrsignif}
\alias{ggrsignif}
\title{Identify Significant Growth Periods in a Differentiated Monte Carlo Density Model}
\usage{
ggrsignif(MCD, probs = c(0.05, 0.9), threshold = 0)
}
\arguments{
\item{MCD}{An object of class \code{'diffMCd'}, representing a differentiated Monte Carlo density model.}

\item{probs}{Numeric vector. The probabilities defining the bounds for significance. Defaults to \code{c(0.05, 0.9)}.}

\item{threshold}{Numeric. The threshold value for identifying significant positive or negative growth. Defaults to \code{0}.}
}
\value{
A data frame with the following columns:
\itemize{
\item \code{year}: The time steps from the input \code{'diffMCd'} object.
\item \code{sig_high}: Logical vector indicating years of significant positive growth.
\item \code{sig_low}: Logical vector indicating years of significant negative growth.
}
The output is assigned the class \code{'ggr_sig'}.
}
\description{
Detects years with significant positive or negative growth in a differentiated Monte Carlo density model (\code{'diffMCd'}).
}
\details{
The function identifies years where the growth rate is significantly higher or lower than the specified threshold. It uses the quantile bounds defined by \code{probs} to determine significance. Positive significance (\code{sig_high}) occurs when the upper quantile exceeds the threshold, while negative significance (\code{sig_low}) occurs when the lower quantile is below the threshold.

NA values are treated as non-significant (\code{FALSE}).
}
\examples{
dates <- data.frame(BP = c(4840, 4885, 4739, 4826, 4610), sd = c(45, 50, 27, 24, 31))
denmod <- MCdensity(dates)
diff_denmod <- ggr(denmod)
plot(denmod)
polygon(ggrsignif(diff_denmod))

}
\seealso{
\code{\link{ggr}}, \code{\link{plot.diffMCd}},  \code{\link{polygon.ggr_sig}}
}
