% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hdr.R
\name{hdr}
\alias{hdr}
\title{Find the interval of the Highest Density Region (hdr) for a Set of Dates}
\usage{
hdr(date, prob = 0.95)
}
\arguments{
\item{date}{A matrix or data frame containing two columns: the first column represents the dates, and the second column represents their corresponding densities.}

\item{prob}{The probability threshold for the region (default is 0.95).}
}
\value{
A list of dates representing the highest density interval(s), along with their corresponding densities.
}
\description{
This function calculates the highest density region (hdr) for a radiocarbon determination.
}
\details{
The hdr represents the range of dates that encompass the specified probability regions in the probability distribution of the age-calibrated dates.
}
\examples{
# Calculate the 95\% hdr for 5310±45 BP 
hdr(rowcal(5310, 45))

# alculate the 68\% hdr for 5310±45 BP 
hdr(rowcal(5310, 45), prob = 0.68)

References
This is a modified version of the same function in the package `Bchron` see Haslett, J. and Parnell, A. 2008 A simple monotone process with application to radiocarbon-dated depth chronologies. Journal of the Royal Statistical Society: Series C (Applied Statistics) 57, 399-418.
}
