% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{intcal}
\alias{intcal}
\title{IntCal20 radiocarbon age calibration curve for the Northern hemisphere.}
\format{
A data.frame with the following fields:
\describe{
\item{\code{calBP}}{Calendar age in years before AD 1950}
\item{\code{14Cage}}{Radiocarbon age in 14C years BP}
\item{\code{Error}}{Radiocarbon age error}
\item{\code{Delta14C}}{Delta 14C value}
\item{\code{Sigma}}{Error of the delta 14C value}
}
}
\source{
https://intcal.org/curves/intcal20.14c
}
\usage{
intcal
}
\description{
IntCal20 radiocarbon age calibration curve for the Northern hemisphere.
}
\references{
Reimer, P. J., Austin, W. E. N., Bard, E., Bayliss, A., Blackwell, P. G., Ramsey, C. B., et al. (2020). The IntCal20 Northern Hemisphere Radiocarbon Age Calibration Curve (0–55 cal kBP). Radiocarbon, 62(4), 725–757. https://doi.org/10.1017/RDC.2020.41
}
\keyword{datasets}
