% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kBP.R
\name{kBP}
\alias{kBP}
\title{Convert Density Models to kyr BP}
\usage{
kBP(X)
}
\arguments{
\item{X}{A data frame or matrix, typically of class \code{rowyear}, \code{MCd} or \code{diffMCd}, where the first column represents calendar years (e.g., BC/AD).}
}
\value{
A modified version of \code{X} with the first column (time axis) converted to thousands of years BP (kyr BP).
}
\description{
This function converts the time axis of a Monte Carlo density model (or similar data)
from calendar years to thousands of years Before Present (kyr BP).
}
\details{
The function transforms the time axis from calendar years (BC/AD) to thousands of years BP,
where BP is calculated relative to 1950 CE. The result is scaled by dividing by 1000 to express the time in kyr BP.

If \code{X} is not of class \code{MCd} or \code{diffMCd}, the function issues a warning but will attempt the transformation.
}
\examples{
# Example with an MCd object
date<-kBP(rowcal(22000,145))
plot(date, xlab='kyr BP', xlim=c(27,25.5))
}
\seealso{
\code{\link{rowcal}}
}
