% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/log.MCd.R
\name{log.MCd}
\alias{log.MCd}
\title{Log Transform the Y-Axis of an MCd Object}
\usage{
\method{log}{MCd}(x)
}
\arguments{
\item{x}{A data frame or matrix of class \code{MCd}, where the first column represents the time axis,
and the remaining columns contain density values.}
}
\value{
A transformed \code{MCd} object with the density values (columns 2 onwards) replaced by their natural logarithms.
}
\description{
This function applies a natural logarithm transformation to the density values (y-axis)
of a Monte Carlo density model (MCd) object. This is sometimes useful for the models produced by the \link{MCmoving} function
}
\details{
The function modifies the input object by applying a natural logarithm to all density values.
The class of the object remains \code{MCd}. This transformation is useful for visualizing density
models with a wide range of values or highlighting relative differences on a logarithmic scale.
}
\examples{
# Example with an MCd object
years <- seq(-5000, 2000, by = 100)
densities <- matrix(runif(100 * length(years), min = 0.01, max = 1), nrow = length(years))
densities <- cbind(years, densities)
class(densities) <- "MCd"
log_transformed <- log.MCd(densities)
head(log_transformed)

}
