% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{marine}
\alias{marine}
\title{Marine20 radiocarbon age calibration curve for the marine environment.}
\format{
A data.frame with the following fields:
\describe{
\item{\code{calBP}}{Calendar age in years before AD 1950}
\item{\code{14Cage}}{Radiocarbon age in 14C years BP}
\item{\code{Error}}{Radiocarbon age error}
\item{\code{Delta14C}}{Delta 14C value}
\item{\code{Sigma}}{Error of the delta 14C value}
}
}
\source{
https://intcal.org/curves/marine20.14c
}
\usage{
marine
}
\description{
Marine20 radiocarbon age calibration curve for the marine environment.
}
\references{
Heaton, T. J., Köhler, P., Butzin, M., Bard, E., Reimer, R. W., Austin, W. E. N., et al. (2020). Marine20—The Marine Radiocarbon Age Calibration Curve (0–55,000 cal BP). Radiocarbon, 62(4), 779–820. https://doi.org/10.1017/RDC.2020.68
}
\keyword{datasets}
