% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MCdensity.R
\name{MCdensity}
\alias{MCdensity}
\title{Calculate Gaussian Kernel Density Estimation (KDE) for a list of age-calibrated probability distributions}
\usage{
MCdensity(
  L = NULL,
  dl = CLIP(),
  default_calcurve = "intcal",
  N = 100,
  bw = 30,
  boot = FALSE
)
}
\arguments{
\item{L}{List of data frames where each element represents a two-column probability distribution. The first column is time, and the second column is probability. If NULL, the function will use the input dataframe dl.}

\item{dl}{A data frame with two or three columns: radiocarbon date BP, sigma, and optionally a calibration curve. Not used if L is specified. Default is to read from the clipboard.}

\item{default_calcurve}{The default calibration curve to use if none is specified in the input data frame. Default is 'intcal'. Ignored if L is specified.}

\item{N}{Number of Monte Carlo iterations. Default is 100.}

\item{bw}{bw The smoothing bandwidth for density estimation. Default is 30 years. Can also be a character string for automatic selection, ee ‘bw.nrd’.}

\item{boot}{Logical. If TRUE, the function will perform a bootstrap resampling of the input data. Default is FALSE.}
}
\value{
A matrix containing the Gaussian Kernel Density Estimation for each Monte Carlo iteration, allowing summary statistics.
}
\description{
This function calculates the Gaussian Kernel Density Estimation (KDE) for a list of two-column probability distributions, where each element represents a probability distribution (time, probability).
}
\details{
The MCdensity function computes a density estimate for a set of radiocarbon calibration curves or similar temporal data, either through bootstrapping or direct sampling.
It provides a density matrix where each column corresponds to a Monte Carlo iteration.
The \code{MCdensity} function generates uncertainty-aware probability density estimates for calibrated radiocarbon dates or other temporal data.
It can handle both deterministic sampling and bootstrapping to model variability in the input data. For sparse datasets, users should set \code{boot} to \code{TRUE} to ensure the model produced is not biased by unevenness in the input data.
For large dataets this is unlikely to make a difference to the resulting density model, so for computational efficiency the direct sampling approach employed when \code{boot=FALSE} can be used.
}
\examples{
# Calculate Gaussian KDE for a mixture of calibrated and calendar dates
L1 <- rowcal(c(5310,5200,-3900, 5100), c(34,43,0.5, 30),
             c('intcal','intcal','calcal','intcal'))
L2 <- rowunif(c(-4000, -4100, -4050), c(-3800, -3900, -3850))
denmod <- MCdensity(c(L1,L2))
plot(denmod)

}
\references{
McLaughlin, T.R. 2019. On applications of space-time modelling with open-source 14C age calibration. Journal of Archaeological Method and Theory 26, 479–501. https://doi.org/10.1007/s10816-018-9381-3
}
\seealso{
\code{\link{MCdensity}} \code{\link{plot.MCd}} \code{\link{density}} \code{\link{bw.nrd}}
}
\author{
T. Rowan McLaughlin
}
