% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MChist.R
\name{MChist}
\alias{MChist}
\title{Bootstrapped Histogram for Radiocarbon Data}
\usage{
MChist(
  datelist,
  bw = 100,
  Nboot = 100,
  plot = TRUE,
  xlab = "Cal. BC/AD",
  col = NA,
  ecol = "black",
  elwd = 1,
  elty = 1,
  ...
)
}
\arguments{
\item{datelist}{A list of 'dates' i.e. two-column matrices such as the output of \code{rowcal} or \code{rowunif}. The first column represents calibrated date, and the second column represents probability.}

\item{bw}{Numeric. The bin width for the histogram. Defaults to \code{100}.}

\item{Nboot}{Integer. Number of bootstrap iterations to perform. Defaults to \code{100}.}

\item{plot}{Logical. If \code{TRUE}, plots the histogram. If \code{FALSE}, returns the histogram object. Defaults to \code{TRUE}.}

\item{xlab}{Character. Label for the x-axis. Defaults to \code{'Cal. BC/AD'}.}

\item{col}{Character. Color of the histogram bars. Defaults to \code{NA}.}

\item{ecol}{Character. Color of error bars in the plot. Defaults to \code{1} (black).}

\item{elwd}{Numeric. Line width for error bars. Defaults to \code{1} (black).}

\item{elty}{Numeric. Line type for error bars. Defaults to \code{1} (black).}

\item{...}{Additional arguments passed to the \code{\link{plot.MChistogram}} function.}
}
\value{
If \code{plot = TRUE}, a histogram plot is generated. If \code{plot = FALSE}, an object of class \code{'MChistogram'} is returned.
The object includes the following elements:
\itemize{
\item \code{mids}: Midpoints of the histogram bins.
\item \code{counts}: Mean counts per bin across bootstraps.
\item \code{bootstrapped_counts}: Matrix of bootstrapped counts for each bin.
}
}
\description{
Generates a bootstrapped histogram for radiocarbon data, incorporating calibration uncertainty into the distribution.
The histogram bins are determined by a specified bin width, and the output can be either a plot or the histogram object.
}
\details{
The function creates a histogram of calibrated radiocarbon dates with uncertainty incorporated through Monte Carlo resampling.
The calibration uncertainty is captured by bootstrapping (\code{Nboot} iterations) and represented as error bars if plotted.
}
\examples{
\dontrun{
# Example data
datelist <- rowcal(c(4000, 4200, 4400, 4300, 5000),
                  c(50, 60, 40, 25,25),
                  c('intcal', 'intcal', 'intcal', 'intcal','marine')
)

# Create and plot a bootstrapped histogram
MChist(datelist, bw = 200, Nboot = 50, ecol = 'red')

# Create the histogram object without plotting
hist_obj <- MChist(datelist, plot = FALSE)
}

}
\references{
McLaughlin, T. R. 2020. An archaeology of Ireland for the Information Age. Emania 25, 7–30. https://www.researchgate.net/publication/347463263_An_archaeology_of_Ireland_for_the_Information_Age
}
\seealso{
\code{\link{plot.MChistogram}} \code{\link{hist}}
}
\author{
T. Rowan McLaughlin
}
