% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MCmix.R
\name{MCmix}
\alias{MCmix}
\title{Convert a Three-column table of Dates into a vector of One Sample per Date}
\usage{
MCmix(L)
}
\arguments{
\item{L}{A three-column list containing radiocarbon determinations or calendar dates. Columns are: date, error, and cc (calibration curve name).}
}
\value{
A vector containing one sample per date, incorporating both radiocarbon determinations and calendar dates.
}
\description{
This function converts a three-column list of dates into a list of one sample per date, incorporating radiocarbon determinations and calendar dates.
}
\details{
This function is a wrapper for \code{MCsam.rowyears} processes each row of the input and generates one sample per date based on the provided radiocarbon determinations or calendar dates.
For calendar dates (cc = 'calcal'), the function samples from the calendar dates using the \code{calsam} function.
For radiocarbon determinations, the function samples from the calibration curve using the \code{rowcalsam} function.
}
\examples{
L <- data.frame(date = c(-3800, 4990, 5300), error = c(10, 30, 50),
      cc = c('calcal', 'intcal', 'marine'))
MCmix(L)

}
\seealso{
\code{\link{MCsam.rowyears}}
}
