% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MCr.as.MCd.R
\name{MCr.as.MCd}
\alias{MCr.as.MCd}
\title{Convert Monte Carlo Running Median Results to Monte Carlo Density Model}
\usage{
MCr.as.MCd(x)
}
\arguments{
\item{x}{A list of matrices of class \code{MCr}. Each matrix should have two columns:
\itemize{
\item The first column: Simulated chronological data.
\item The second column: Corresponding running median values.
}}
}
\value{
A matrix of class \code{MCd}, where:
\itemize{
\item The first column contains the regularly spaced time grid.
\item The remaining columns contain interpolated running median values for each simulation.
}
}
\description{
This function converts a list of Monte Carlo running median (\code{MCr}) results
into a Monte Carlo density model (\code{MCd}) object by interpolating the data
to a regular time grid.
}
\details{
The function first determines the minimum and maximum chronological range across all
simulations and generates a regular time grid. It then interpolates the running median
results of each simulation to this grid using linear interpolation. Rows with all \code{NA}
values (excluding the time grid) are removed from the final output.
}
\section{Example data}{
x0 <- c(-5010, -5110, -5350) # from
x1 <- c(-5000, -5100, -5200) # to
y <- c(0.2, 0.7, 0.9)
}

\section{Perform Monte Carlo simulation}{
result <- MCrunmed(x0, x1, y, k = 3, N = 10, blur = 0.1)
plot(MCr.as.MCd(result), ylab='Response')
}

\seealso{
\code{\link{MCrunmed}} \code{\link{plot.MCd}}
}
