% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MCsam.R
\name{MCsam}
\alias{MCsam}
\alias{MCsam.default}
\title{Convert a Two-column List of Uncalibrated Dates into a List of One Sample per Date
This function converts a two-column list of uncalibrated radiocarbon dates into a list of one sample per date.}
\usage{
MCsam(x, ...)
}
\arguments{
\item{...}{Additional arguments to be passed to the \code{rowcalsam} function.}

\item{L}{A two-column list containing uncalibrated radiocarbon dates. Columns are: BP (Before Present) and SD (Standard Deviation).}
}
\value{
A vector containing one sample per radiocarbon date.
}
\description{
Convert a Two-column List of Uncalibrated Dates into a List of One Sample per Date
This function converts a two-column list of uncalibrated radiocarbon dates into a list of one sample per date.
}
\details{
This function processes each row of the input list and generates one sample per uncalibrated radiocarbon date using the \code{rowcalsam} function.
}
\examples{
L <- data.frame(BP = c(5310, 5200, 5030), SD = c(35, 41, 23))
MCsam(L)

}
\author{
T. Rowan McLaughlin
}
