% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MCsam.rowyears.R
\name{MCsam.rowyears}
\alias{MCsam.rowyears}
\title{Convert a List of Calibrated Dates Into a Vector of One Sample per Date}
\usage{
\method{MCsam}{rowyears}(L, n = 1)
}
\arguments{
\item{L}{A list containing calibrated or calendar dates, each element being a matrix with two columns: calendar dates and corresponding probabilities.}
}
\value{
A vector containing one sample per date.
}
\description{
This function converts a list of calibrated radiocarbon dates into a vector of one sample per date.
}
\details{
This function processes each element of the input list and generates one sample per calibrated radiocarbon date.
Note that there is no check that the elements are all using the same calendar.
}
\examples{
L <- rowcal(c(5310, 5200, -3900),c(34,43,0.5),c('intcal','intcal','calcal'))
MCsam(L)
}
\author{
T. Rowan McLaughlin
}
