% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mean.diffMCd.R
\name{mean.diffMCd}
\alias{mean.diffMCd}
\title{Calculate Mean Value of a Growth Rate Model}
\usage{
\method{mean}{diffMCd}(A, at = NULL, from = NULL, to = NULL)
}
\arguments{
\item{A}{Any differentiated \code{MCd} denisty model calculated using \code{ggr}}

\item{at}{Point to calculate the mean density for.}

\item{from}{Starting point for calculating the mean (default is the minimum value in the density model).}

\item{to}{Ending point for calculating the mean (default is the maximum value in the density model).}
}
\value{
The mean value of the density model.
}
\description{
This function calculates the mean value of a density growth rate model, optionally between two points.
}
\details{
This function calculates the mean value of a density model. If \code{at} is specified the mean value returned it the year of the density model closest to that value. Otherwise, the mean density between \code{from} and \code{to} is returned, if either is not specified then the start and /or endpoints of the density model are used in their place.
}
\examples{
# Calculate the mean value of a density model at 4000 cal. BC
mean(ggr(denmod), at=-4000)

# Calculate the mean value of a density model between two points
mean(ggr(denmod), from = -1200, to = -1000)

}
