% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mean.MCd.R
\name{mean.MCd}
\alias{mean.MCd}
\title{Calculate Mean Value of a Density Model}
\usage{
\method{mean}{MCd}(A, at = NULL, from = NULL, to = NULL)
}
\arguments{
\item{A}{Any \code{MCd} object such as those produced by \code{MCdensity}, \code{phasedensity} or \code{bootstrap_density}}

\item{at}{Point to calculate the mean density for.}

\item{from}{Starting point for calculating the mean (default is the minimum value in the density model).}

\item{to}{Ending point for calculating the mean (default is the maximum value in the density model).}
}
\value{
The mean value of the density model.
}
\description{
This function calculates the mean value of a density model, optionally between two points.
}
\details{
This function calculates the mean value of a density model. If \code{at} is specified the mean value returned it the year of the density model closest to that value. Otherwise, the mean density between \code{from} and \code{to} is returned, if either is not specified then the start and /or endpoints of the density model are used in their place.
}
\examples{
# Calculate the mean value of a density model at 4000 cal. BC
mean.MCd(denmod, at=-4000)

# Calculate the mean value of a density model between two points
mean.MCd(denmod, from = -1200, to = -1000)

}
