% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/median.MCd.R
\name{median.MCd}
\alias{median.MCd}
\title{Calculate Median of a Density Model}
\usage{
\method{median}{MCd}(x)
}
\arguments{
\item{x}{Density model data, typically output from functions like MCdensity or phasedensity.}
}
\value{
A matrix containing the x-values and corresponding medians of the density model.
}
\description{
This function calculates the median of a density model.
}
\examples{
# Calculate median of density model
median_model <- median.MCd(MCdensity_output)

}
