% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/median.rowyear.R
\name{median.rowyear}
\alias{median.rowyear}
\title{Find the median date for a \code{rowyear} object.}
\usage{
\method{median}{rowyear}(mat)
}
\arguments{
\item{L}{A two-column matix typically of \code{rowyear} fomat; columns are: date, probablitiy.}
}
\value{
The 'median' date of the \code{rowyear} object.
}
\description{
This function returns the 'median' date by calculated by finding the row that corresponds to the 50th percentile of the cumulative probability.
}
\examples{

median(rowcal(5030, 50, 'intcal'))
median(rowcal(5030, 50, 'marine')
median(rowcal(-4000, 50, 'calcal')) # Should really be -4000 !

}
\seealso{
\code{\link{rowcal}} \code{\link{findmedian}} \code{\link{findmode}} \code{\link{findwm}}
}
\author{
T. Rowan McLaughlin
}
