% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/median.SPDboot.R
\name{median.SPDboot}
\alias{median.SPDboot}
\title{Find the Median of Summed Probability Distribution (SPD) Bootstraps}
\usage{
\method{median}{SPDboot}(SPDb)
}
\arguments{
\item{SPDb}{A matrix containing the bootstrapped SPDs, where each column represents a bootstrap iteration.}
}
\value{
A list containing two elements: \code{x} (the calendar years) and \code{y} (the median values).
}
\description{
This function calculates the median of a set of Summed Probability Distribution (SPD) bootstraps.
}
\details{
This function calculates the median of a set of SPD bootstraps. It applies the \code{median} function to each row of the bootstrapped SPD matrix to determine the median values for each calendar year, i.e. the 'population proxy' of Fernández-López de Pablo et al 2019.
}
\examples{
# Plot bootstrapped confidence intervals for a SPD 
dates <- data.frame(BP = c(4840, 4885, 4739, 4840, 4885, 4239, 4840, 4845, 4739, 4826, 4610), sd = c(45, 50,60,30, 27, 30, 33, 45, 23, 24, 31))
S<-rowcalsum(dates)
Sboot<-SPDboot(S)
plot(Sboot, col = "#8888FF20")
population_proxy <- median(Sboot)
lines(population_proxy, col="#5050FF")

}
\references{
Fernández-López de Pablo, J., Gutiérres-Roig, M.,Gómez-Puche, M., Silva, F., McLaughlin, R., and Lozano, S. 2019. Palaeo-demographic modelling supports a population bottleneck during the Pleistocene-Holocene transition in Iberia. Nature Communications 10, 1872. http://doi.org/10.1038/s41467-019-09833-3
\code{\link{SPDboot}} \link{'plot.SPDboot'}
}
