% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mixcurves.R
\name{mixcurves}
\alias{mixcurves}
\title{Mix Curves and Apply Local Delta-R Corrections}
\usage{
mixcurves(
  mix = 0.5,
  uncert = 0,
  c1 = intcal,
  c2 = marine,
  delR = 0,
  error_delR = 0
)
}
\arguments{
\item{mix}{Mixing proportion between the two curves. Default is 0.5.}

\item{uncert}{Uncertainty value. Default is 0.}

\item{c1}{Calibration curve 1, typically in calibration years BP (Before Present).}

\item{c2}{Calibration curve 2, typically in calibration years BP (Before Present).}

\item{delR}{Delta-R value for local delta-R correction. Default is 0.}

\item{error_delR}{Error associated with delta-R correction. Default is 0.}
}
\value{
A matrix containing the mixed curves and applied corrections.
}
\description{
This function mixes two calibration curves and applies local delta-R corrections if required.
}
\details{
This function interpolates the second calibration curve to match the calendar years of the first calibration curve.
It then applies delta-R correction if required and performs the mixing of the curves based on the mixing proportion.
The output matrix contains the mixed curves along with the applied corrections.
}
\examples{
# Mixing calibration curves
mixcurves(mix = 0.5, uncert = 0, c1 = intcal, c2 = marine, delR = 0, error_delR = 0)

}
