% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mixdensity.R
\name{mixdensity}
\alias{mixdensity}
\title{Calculate Gaussian Kernel Density Estimation (KDE) for a set of radiocarbon determinations using various calibration curves, or calendar age estimates}
\usage{
mixdensity(
  DATA = CLIP(),
  N = 100,
  perm_runs = 1,
  perm_size = 1,
  col = rgb(0, 0, 0, 0.01),
  plot.new = FALSE,
  add = FALSE,
  bw = 30,
  ...
)
}
\arguments{
\item{DATA}{A 3-column data frame containing dates, errors, and calibration curve information. By default, it reads data from the clipboard. Columns should be: date, error, curve.}

\item{N}{Number of Monte Carlo iterations.}

\item{perm_runs}{Number of permutations of the data. Default is 1.}

\item{perm_size}{Proportion of data to be used for each permutation. Default is 1.}

\item{col}{Colour of the lines in the optional density 'ghost' plot. Default is transparent black.}

\item{plot.new}{Logical indicating whether to open a new plot window. Default is FALSE.}

\item{add}{Logical indicating whether to add the density plot to an existing plot. Default is FALSE.}

\item{bw}{The smoothing bandwidth for density estimation. Default is 30 years. Can also be a character string for automatic selection, ee ‘bw.nrd’.}

\item{...}{Additional arguments to be passed to the \code{density} function.}
}
\value{
A matrix of type 'MCd' containing the Gaussian Kernel Density Estimation for each Monte Carlo iteration, allowing summary statistics.
}
\description{
This function calculates the Gaussian Kernel Density Estimation (KDE) model of a set of dates using a Monte Carlo method. The input data is a table contain radiocarbon dates with errors, associated with different calibration curves, or calendar age estimates.
}
\details{
This function performs Gaussian Kernel Density Estimation using a Monte Carlo method for a set of dates with errors and associated calibration curve information.
It calculates the density estimate for each Monte Carlo iteration and returns a matrix containing the density estimates for further analysis.
By default, the densities are calculated at 512 points in time. This can be changed by editing the source, but it needs to be a power of 2 (e.g., 1024, 2048, etc.).
}
\examples{
# Calculate Gaussian KDE for a mixed date set with default parameters
mixdensity()

# Calculate Gaussian KDE for a mixed date set with custom parameters
mixdensity(N = 1000, perm_runs = 5, perm_size = 0.8, col = "blue", plot.new = TRUE, add = TRUE, bw = 20)

}
\references{
McLaughlin, T.R. 2019. On applications of space-time modelling with open-source 14C age calibration. Journal of Archaeological Method and Theory 26, 479–501. https://doi.org/10.1007/s10816-018-9381-3
}
\seealso{
\code{\link{density}} \link{'MCdensity'} \link{'phasedensity'} \code{\link{bw.nrd}} \code{\link{plot.MCd}}
}
