% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phasedensity.R
\name{phasedensity}
\alias{phasedensity}
\title{Calculate Density Model for Radiocarbon Dates using Phase Binning}
\usage{
phasedensity(
  DATA = CLIP(),
  N = 100,
  perm_runs = 1,
  perm_size = 1,
  col = rgb(0, 0, 0, 0.01),
  plot.new = FALSE,
  add = FALSE,
  bw = 30,
  h = 30,
  method = "sample",
  shuffle = TRUE,
  ...
)
}
\arguments{
\item{DATA}{Data frame containing radiocarbon dates, labeled with some form of site name or code. Columns should be: siteid, date, error, \link{calcurve}.}

\item{N}{Number of Monte Carlo runs. Default is 100.}

\item{perm_runs}{Number of permutations of the data. Default is 1.}

\item{perm_size}{Size of permutations as a fraction of the original data size. Default is 1, i.e. 100\%.}

\item{col}{Color for plotting density curves. Default is transparent black.}

\item{plot.new}{Logical indicating whether to create a new plot. Defaults to FALSE.}

\item{add}{Logical indicating whether to add density curves to an existing plot. Defaults to FALSE.}

\item{bw}{The smoothing bandwidth for density estimation. Default is 30 years. Can also be a character string for automatic selection, ee ‘bw.nrd’.}

\item{h}{Height parameter for dendrogram cutting in phase binning. Default is 30 years.}

\item{method}{Method for binning dates in phase binning: 'median' to bin using median dates or 'sample' for Monte Carlo draws.}

\item{shuffle}{Logical indicating whether to shuffle the input data before processing. Defaults to FALSE.}

\item{...}{Additional arguments to be passed to the \code{plot} function if plot.new==TRUE.}
}
\value{
A matrix of type 'MCd' containing the density model for the radiocarbon dates.
}
\description{
This function calculates the Gaussian Kernel Density Estimation (KDE) model  for a set of radiocarbon dates using phase binning to select one date per phase.
}
\details{
This function calculates a density model for a set of radiocarbon dates in the same manner as \code{mixdensity}. However, in addition, the data are parsed using hierarchal clustering.
It first performs phase binning of radiocarbon dates using the \code{phasesam} function and then calculates the density model using the selected dates using either \code{mixdensity}, or \code{MCdensity} for cases where only one calibration curve is used, as \code{MCdensity} is slightly faster than \code{mixdensity}.
}
\examples{
# Calculate density model with default parameters
phasedensity()

# Calculate density model with custom parameters
phasedensity(DATA = my_data, N = 200, perm_runs = 3, perm_size = 0.5, col = "blue", plot.new = TRUE, add = FALSE, bw = 25, h = 35, method = "sample", shuffle = TRUE)

}
\references{
McLaughlin, T.R. 2019. On applications of space-time modelling with open-source 14C age calibration. Journal of Archaeological Method and Theory 26, 479–501. https://doi.org/10.1007/s10816-018-9381-3
For similar analysis with respect to clustering see Crema ER, Bevan A. 2021. Inference From Large Sets of Radiocarbon Dates: Software and Methods. Radiocarbon 63(1):23-39. doi:10.1017/RDC.2020.95
}
\seealso{
\code{\link{density}} \link{'MCdensity'} \link{'mixdensity'} \code{\link{phasesam}} \code{\link{bw.nrd}} \code{\link{plot.MCd}}
}
