% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phaser.R
\name{phaser}
\alias{phaser}
\title{Perform Phase Binning of Radiocarbon Dates}
\usage{
phaser(DATA = CLIP(), h = 30)
}
\arguments{
\item{DATA}{A 3- or 4-column data frame containing radiocarbon dates. Columns should be: siteid, date, error, \link{calcurve}.}

\item{h}{Height parameter for dendrogram cutting. Default is 30 years.}
}
\value{
A data frame containing the input data with an additional column indicating the assigned time bin for each date.
}
\description{
This function performs phase binning of radiocarbon determinations using hierarchical clustering, assigning each date to a phase based the median value of the calibrated dates.
}
\details{
This function performs phase binning of radiocarbon dates by assigning each date to a phase based on median dates.
It calculates the median date for each site and then uses hierarchical clustering to group the dates into phases.
}
\examples{
# Perform phase binning with default parameters
phaser()

# Perform phase binning with custom height parameter
phaser(DATA = my_data, h = 25)

}
