% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phasesam.R
\name{phasesam}
\alias{phasesam}
\title{Select one date per phase in a set of radiocarbon determinations, using hierarchical clustering to identify phases}
\usage{
phasesam(DATA = CLIP(), h = 30, method = "median", shuffle = TRUE)
}
\arguments{
\item{DATA}{A 3- or 4-column dataframe containing radiocarbon dates. Columns should be: siteid, date, error, \link{calcurve}.}

\item{h}{Height parameter for dendrogram cutting. Default is 30 years for 30-year phases at each site.}

\item{method}{Method for binning dates: 'median' to bin using median dates or 'sample' for Monte Carlo draws.}

\item{shuffle}{Logical indicating whether to shuffle the input data before processing.}
}
\value{
A data frame containing the input data thinned to one date per phase.
}
\description{
This function performs phase sampling of radiocarbon dates, binning them into unique phases using hierarchical clustering based on either median dates or Monte Carlo draws.
}
\details{
This function performs phase sampling of radiocarbon dates by binning them into phases based on either median dates or Monte Carlo draws.
If the \code{shuffle} parameter is TRUE, the input data will be shuffled before processing. Otherwise the first date per phase in the list will always be selected for analysis.
If any NA values are introduced during the binning process, the function will repeat the Monte Carlo sampling until all NAs are resolved.
}
\examples{
# Perform phase sampling with default parameters
phasesam()

# Perform phase sampling with custom parameters
phasesam(DATA = my_data, h = 25, method = 'sample', shuffle = FALSE)

}
\references{
For a similar analysis see Crema ER, Bevan A. 2021. Inference From Large Sets of Radiocarbon Dates: Software and Methods. Radiocarbon 63(1):23-39. doi:10.1017/RDC.2020.95
}
