% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.c14sum.R
\name{plot.c14sum}
\alias{plot.c14sum}
\title{Plot a Summed Probability Distribution (SPD) of Radiocarbon Dates}
\usage{
\method{plot}{c14sum}(
  S,
  add = FALSE,
  offset = 0,
  xlab = "Automatic",
  ylab = "Summed prob.",
  type = "f",
  col = rgb(0.8, 0, 0, 0.8),
  lcol = col,
  lty = 1,
  lwd = 1,
  sm = 1,
  ...
)
}
\arguments{
\item{S}{A list containing the summed probability distribution, typically generated by the \code{rowcalsum} function.}

\item{add}{Logical value indicating whether to add the plot to an existing plot (default is FALSE).}

\item{offset}{A value to add as an offset to the y-axis (default is 0).}

\item{xlab}{Label for the x-axis (default 'Automatic' will attempt to label the x-axis given the date range of the SPD).}

\item{ylab}{Label for the y-axis (default is 'Summed prob.').}

\item{type}{The type of plot: 'f' for filled polygons, 'l' for lines or 'b' for both.}

\item{col}{Color for the plot (default is a translucent red).}

\item{lcol}{Color for lines if type='b' or 'l' (default is to take the value of col).}

\item{lty}{Line type for the plot (default is 1).}

\item{lwd}{Line width for the plot (default is 1).}

\item{sm}{a smoothing parameter; the size of a moving average filter (default is 1). Higher values result in more smoothing.}

\item{...}{Additional graphical parameters passed to the \code{plot} function.}
}
\description{
This function plots a summed probability distribution (SPD) for a given radiocarbon date dataset.
}
\details{
This function plots a summed probability distribution for a given radiocarbon date dataset. It can create a new plot or add to an existing plot. The plot type can be either filled polygons or lines.
}
\examples{
# Plot a summed probability distribution with filled polygons
dates <- data.frame(BP = c(4840, 4885, 4739, 4826, 4610), sd = c(45, 50, 27, 24, 31))
spd<-rowcalsum(dates)
plot.c14sum(spd)

# Plot a summed probability distribution with lines, smoothed and unsmoothed
plot.c14sum(spd, type = 'l', col = 1)
plot.c14sum(spd, type = 'l', col = 1, lwd = 2, sm=4)

[`rowcalsum`]
}
