% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.diffMCd.R
\name{plot.diffMCd}
\alias{plot.diffMCd}
\title{Plot Geometric Growth Rate ('diffMCd' Object)}
\usage{
\method{plot}{diffMCd}(MCD, ..., endzero = FALSE, ylab = "Growth rate / \%", ylim = c(-2, 2))
}
\arguments{
\item{MCD}{An object of class \code{'diffMCd'}, typically created using the \code{\link{ggr}} function.}

\item{...}{Additional arguments passed to \code{\link{plot.MCd}} for customization.}

\item{endzero}{Logical. If \code{TRUE}, adds rows with zero growth rates at the beginning and end of the object to pad the data. Defaults to \code{FALSE}.}

\item{ylab}{Character. Label for the y-axis. Defaults to "Growth rate / \%".}

\item{ylim}{Numeric vector. Range for the y-axis. Defaults to \code{c(-2, 2)}.}
}
\value{
A plot is produced on the active graphics device.
}
\description{
A method for plotting objects of class \code{'diffMCd'}, which represent geometric growth rates calculated from an 'MCd' object.
}
\details{
This function visualizes the growth rates contained in a \code{'diffMCd'} object. By default, it filters out rows where the mean density is below a threshold (1e-7) or contains non-finite values.
If \code{endzero} is set to \code{TRUE}, zero-growth rows are appended at the beginning and end of the object for visual continuity.

The actual plotting is delegated to the \code{\link{plot.MCd}} function, with any additional arguments passed through \code{...}.
}
\examples{
\dontrun{
data(BIRE)
IrD<-MCdensity(BIRE[which(BIRE$Where=='Ireland'),2:3])
par(mfrow=c(2,1))
plot(IrD, xlim=c(-6300,-2500))
plot(ggr(IrD),xlim=c(-6300,-2500))
}

}
\seealso{
\code{\link{ggr}}, \code{\link{plot.MCd}}
}
