% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.MCd.R
\name{plot.MCd}
\alias{plot.MCd}
\title{Plot Density Models}
\usage{
\method{plot}{MCd}(
  x,
  add = FALSE,
  col = 1,
  fill = "#00000022",
  scalefactor = 1,
  smax = FALSE,
  ylab = "Density",
  xlab = "Automatic",
  Toffset = 0,
  grid = FALSE,
  lwd = 1,
  lty = 1,
  sigma = 1,
  ...
)
}
\arguments{
\item{x}{Density model data, typically output from functions like MCdensity, mixdensity, or phasedensity.}

\item{add}{Logical indicating whether to add to an existing plot (TRUE) or create a new one (FALSE).}

\item{col}{Color of the density line. Default is black.}

\item{fill}{Fill color for the shaded area (defined by the parameter \code{sigma}). Default is transparent grey.}

\item{scalefactor}{Scaling factor for adjusting the density model.}

\item{smax}{Logical indicating whether to normalize the density to the maximum value.}

\item{ylab}{Label for the y-axis. Default is 'Density'.}

\item{xlab}{Label for the x-axis. If not provided, the function will try to guess the most appropriate label based on the first column of \code{x}.}

\item{Toffset}{Time offset to be added to x-axis values.}

\item{grid}{Logical indicating whether to add a grid to the plot.}

\item{lwd}{Line width for the density line.}

\item{lty}{Line type for the density line.}

\item{sigma}{Standard deviation multiplier for the shaded area.}

\item{...}{Additional graphical parameters to be passed to the \code{plot} function.}
}
\value{
A plot displaying the density model.
}
\description{
This function plots density models generated from Monte Carlo simulations of radiocarbon dates
}
\details{
It is intended that a future version of this function will take a list of \code{MCd} objects and plot them with less manual intervention.
}
\examples{
data(BIRE)
plot(IrD, xlim=c(-6300,-2500))

}
