% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.MCd_summary.R
\name{plot.MCd_summary}
\alias{plot.MCd_summary}
\title{Plot Summary Form of Density Models}
\usage{
\method{plot}{MCd_summary}(
  x,
  add = FALSE,
  col = rgb(0, 0, 0.8, 0.5),
  scalefactor = 1,
  xlab = "Automatic",
  grid = TRUE,
  ...
)
}
\arguments{
\item{x}{Density model data, typically output from functions like MCdensity or phasedensity.}

\item{add}{Logical indicating whether to add to an existing plot (TRUE) or create a new one (FALSE).}

\item{col}{Colour of the lines and polygon.}

\item{scalefactor}{Scaling factor for adjusting the density model.}

\item{xlab}{Label for the x-axis. Default is automatic.}

\item{grid}{Logical indicating whether to add a grid to the plot.}

\item{...}{Additional graphical parameters to be passed to plot function.}
}
\value{
A plot displaying the summary form of the density model.
}
\description{
This function plots the summary form of density models generated from Monte Carlo simulations that have already been summarised by \code{summary.MCd}.
}
\details{
This function will be upgraded to match the newer visual syntax of the current version of \code{plot.MCd}
}
\examples{
# Plot summary form of density model
plot.MCd_summary(density_model)
}
\seealso{
\code{\link{plot.MCd}} \code{\link{summary.MCd}}
}
