% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.MChistogram.R
\name{plot.MChistogram}
\alias{plot.MChistogram}
\title{Plot a Bootstrapped MChistogram with Error Bars}
\usage{
\method{plot}{MChistogram}(A, xlab = "Cal. BC/AD", ecol = "black", elwd = 1, elty = 1, ...)
}
\arguments{
\item{A}{An object of class \code{MChistogram}, which contains histogram data including \code{bootstrapped_counts}.}

\item{xlab}{A character string specifying the label for the x-axis. Defaults to \code{'Cal. BC/AD'}.}

\item{ecol}{A character string or color code specifying the color of the error bars. Defaults to \code{'black'}.}

\item{elwd}{A numeric value specifying the line width of the error bars. Defaults to \code{1}.}

\item{elty}{A numeric value specifying the line type of the error bars. Defaults to \code{1}.}

\item{...}{Additional arguments to be passed to the base \code{\link{plot}} function.}
}
\value{
A plot is generated. No value is returned.
}
\description{
This function plots an MChistogram object with bootstrapped mean counts and associated standard deviation (error bars).
It extends the default histogram plot by adding error bars to represent uncertainty.
}
\details{
The function first converts the input object \code{A} to a class of \code{histogram}.
It then calculates the mean counts and standard deviations across bootstrap replicates
(stored in \code{A$bootstrapped_counts}). The histogram is plotted using the default
\code{\link{plot}} function, and error bars representing the standard deviation are added.
}
\examples{
# Example usage
# Assuming A is an MChistogram object
\dontrun{
plot.MChistogram(A)
}
}
\references{
McLaughlin, T. R. 2020. An archaeology of Ireland for the Information Age. Emania 25, 7–30. https://www.researchgate.net/publication/347463263_An_archaeology_of_Ireland_for_the_Information_Age
}
\seealso{
\code{\link{MChist}} \code{\link{plot.histogram}}
}
\author{
T. Rowan McLaughlin
}
