% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.rowyear.R
\name{plot.rowyear}
\alias{plot.rowyear}
\title{Plot a \code{rowyear} probability object}
\usage{
\method{plot}{rowyear}(
  S,
  add = FALSE,
  offset = 0,
  xlab = "Automatic",
  ylab = attr(S, "probs"),
  type = "f",
  col = rgb(0.8, 0, 0, 0.8),
  lcol = col,
  lty = 1,
  lwd = 1,
  sm = 1,
  ...
)
}
\arguments{
\item{S}{A two-column matrix containg the probability distribution to plot.}

\item{add}{Logical value indicating whether to add the plot to an existing plot (default is FALSE).}

\item{offset}{A value to add as an offset to the y-axis (default is 0).}

\item{xlab}{Label for the x-axis (default will attempt to label the x-axis given the \code{probs} attribute of the data).}

\item{ylab}{Label for the y-axis (default is 'Summed prob.').}

\item{type}{The type of plot: 'f' for filled polygons, 'l' for lines or 'b' for both.}

\item{col}{Color for the plot (default is a translucent red).}

\item{lcol}{Color for lines if type='b' or 'l' (default is to take the value of col).}

\item{lty}{Line type for the plot (default is 1).}

\item{lwd}{Line width for the plot (default is 1).}

\item{sm}{a smoothing parameter; the size of a moving average filter (default is 1). Higher values result in more smoothing.}

\item{...}{Additional graphical parameters passed to the \code{plot} function.}
}
\description{
This function plots the pobalility distrubution contained by a \code{rowyear} object, such as a radiocarbon date or summed probability distribution.
}
\details{
This function plots a summed probability distribution for a given radiocarbon date dataset. It can create a new plot or add to an existing plot. The plot type can be either filled polygons or lines.
}
\examples{
# Plot a summed probability distribution with filled polygons
dates <- rowcal(c(4840, 4885, 4739, 4826, 4610), c(45, 50, 27, 24, 31))
spd<-sum(dates)
plot(spd)

# Plot a summed probability distribution with lines, smoothed and unsmoothed
plot.c14sum(spd, type = 'l', col = 1)
plot.c14sum(spd, type = 'l', col = 2, lwd = 2, sm=4,add=TRUE)

[`rowcal`] [`rowunif`] [`sum.rowyears`] [`fill`]
}
