% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.rowyears.R
\name{plot.rowyears}
\alias{plot.rowyears}
\title{Plot a \code{rowyears} (multiple \code{rowyear}) probabilty object}
\usage{
\method{plot}{rowyears}(
  L,
  col = "#AAAAFA",
  lcol = 1,
  blcol = NA,
  type = "b",
  xlab = "Automatic",
  yscale = 10,
  lty = 1,
  lwd = 1,
  add = FALSE,
  ...
)
}
\arguments{
\item{L}{A list of two-column matrices. The first column represents time (e.g., years), and the second column represents density estimates.}

\item{col}{A vector of fill colours for polygons. Recycled if its length is less than the length of \code{L}. Default is \code{'#AAAAFA'}.}

\item{lcol}{A vector of line colours for the plots. Recycled if its length is less than the length of \code{L}. Default is \code{1} (black).}

\item{blcol}{A vector of line colours the base of each . Recycled if its length is less than the length of \code{L}. Default is \code{NA} (no line).}

\item{type}{Character string indicating the type of plot. Options are:
\itemize{
\item \code{'f'}: Filled polygons only.
\item \code{'l'}: Lines only.
\item \code{'b'}: Both lines and filled polygons.
}
Default is \code{'b'}.}

\item{xlab}{A label for the x-axis. Use \code{'Automatic'} for automatic detection of the time scale (e.g., "Cal. BC/AD", "Cal. BP"). Default is \code{'Automatic'}.}

\item{yscale}{A numeric value to scale the density estimates vertically. Default is \code{10}.}

\item{lty}{Line type for the plot. Default is \code{1} (solid line).}

\item{lwd}{Line width for the plot. Default is \code{1}.}

\item{add}{Logical value indicating whether to add the plot to an existing plot. Default is \code{FALSE}.}

\item{...}{Additional arguments passed to \code{\link[graphics]{plot.default}}.}
}
\value{
None. The function is called for its side effect of producing a plot.
}
\description{
Visualizes a list of matrices produced by \code{rowcal} or \code{rowunif} using various plotting styles, including filled polygons, lines, or both.
}
\details{
The function plots each element in \code{L} on the same time axis, stacking them vertically with an offset to distinguish between different elements.
The offset is determined based on the number of elements in \code{L}. The x-axis label is automatically determined unless specified.
}
\examples{
caldates<-c(rowcal(c(4800,4850,5400,-3600),c(30,25,40,75),
            c('intcal','intcal','marine','calcal')))
# add a uniform distribution, NB a single matrix wrapped in a list
# also calculate the SPD and reverse order of list for plotting
dates<-c(caldates, list(rowunif(-3750,-3550)))
spd <- sum.rowyears(dates)
plot.rowyears(c(list(spd),dates),
              col=c(1,'#900','#900','#099','#FFF','#BBB',1),blcol=1,
              xlim=c(-4000,-3400),xaxt='n'); ax()
legend('topright',fill=c('#BBB','#FFF','#099','#900',1),
  c(Uniform','Normal','Marine','IntCal','Sum'))
text(-3900,c(0:6)/6+0.03,lab=c(NA,'4700±30 BP','4750±25 BP','5400±40 BP',
                             '3600±75 cal BC','3750 to 3550 cal BC'))

# simulate 100 radiocarbon dates and plot them with a rainbow color scheme
uncal<-sort(runif(100,3000,5000), decreasing=TRUE)
dates<-rowcal(uncal, runif(100,10,50))
plot(dates, col=rainbow(100, alpha=0.5),lcol='#33333388',xaxt='n')
ax()

}
\seealso{
\code{\link{rowcal}} \code{\link{rowunif}} \code{\link{plot.rowyear}} \code{\link{plot.default}}
}
\author{
T. Rowan McLaughlin
}
