% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.SPDb.R
\name{plot.SPDb}
\alias{plot.SPDb}
\title{Plot Bootstrapped Confidence Intervals for a Summed Probability Distribution (SPD)}
\usage{
\method{plot}{SPDb}(
  SPDb,
  probs = c(0.16, 0.84),
  col = rgb(0.1, 0.7, 0.1, 0.8),
  add = FALSE,
  fill = NA,
  xlab = "Automatic",
  ylab = "Summed prob.",
  ...
)
}
\arguments{
\item{SPDb}{A matrix containing the bootstrapped SPDs (generated by the \code{SPDboot} function), where each column represents a bootstrap iteration.}

\item{probs}{A numeric vector specifying the lower and upper percentiles for the confidence intervals (default is c(0.16, 0.84)).}

\item{col}{Color for the plot (default is a translucent green).}

\item{add}{Logical value indicating whether to add the plot to an existing plot (default is FALSE).}

\item{fill}{Logical value indicating whether to fill the area between the confidence intervals (default is TRUE).}

\item{xlab}{Label for the x-axis (default is 'Automatic').}

\item{ylab}{Label for the y-axis (default is 'Summed prob.').}

\item{...}{Additional graphical parameters passed to the \code{plot} function.}
}
\description{
This function plots bootstrapped confidence intervals for a Summed Probability Distribution (SPD).
}
\details{
This function plots bootstrapped confidence intervals for a given SPD. It calculates the specified percentiles for the bootstrapped SPDs and plots the upper and lower bounds as lines. If \code{fill} is TRUE, it fills the area between the confidence intervals with the specified color.

The function can create a new plot or add to an existing plot, with customizable x-axis and y-axis labels, plot colors, and other graphical parameters.
}
\examples{
# Plot bootstrapped confidence intervals for a SPD
dates <- data.frame(BP = c(4840, 4885, 4739, 4840, 4885, 4239, 4840, 4845, 4739, 4826, 4610),
    sd = c(45, 50,60,30, 27, 30, 33, 45, 23, 24, 31))
S<-rowcalsum(dates)
Sboot<-SPDboot(S)
plot(Sboot, col=NA, fill = "#8888FF20")
lines(median(Sboot), col="#5050FF")

[`SPDboot`] ['median.SPDb'] ['summary.SPDb']
}
