% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_MCd_offsets.R
\name{plot_MCd_offsets}
\alias{plot_MCd_offsets}
\title{Plot Multiple KDE Models with Y-Axis Offsets}
\usage{
plot_MCd_offsets(
  x,
  norm = TRUE,
  add = FALSE,
  col = 1,
  fill = "#e5e5e5e5",
  scalefactor = 1,
  ylab = "Density",
  xlab = "Automatic",
  yaxt = "Automatic",
  yaxt_side = 2,
  Toffset = 0,
  yoffset = 0,
  yaxes = TRUE,
  grid = FALSE,
  lwd = 1,
  lty = 1,
  ...
)
}
\arguments{
\item{x}{Density model data, typically output from functions like MCdensity or phasedensity.}

\item{norm}{Logical indicating whether to normalize the density models. Defaults to TRUE.}

\item{add}{Logical indicating whether to add the plots to an existing plot. Defaults to FALSE.}

\item{col}{Color of the lines. Defaults to 1 (black).}

\item{fill}{Fill color for the area under the curve. Defaults to semi-transparent grey.}

\item{scalefactor}{Scaling factor for adjusting the density models. Defaults to 1.}

\item{ylab}{Label for the y-axis. Defaults to 'Density'.}

\item{xlab}{Label for the x-axis. By defaults the function applies 'Cal. BC'. 'Cal. BC/AD' or 'Cal. AD' as appropriate.}

\item{yaxt}{Specifies the y-axis type. Defaults to 'Automatic'.}

\item{yaxt_side}{Side of the plot to draw the y-axis. Defaults to 2 (left side).}

\item{Toffset}{Offset for the x-axis. Defaults to 0.}

\item{yoffset}{Offset for the y-axis. Defaults to 0.}

\item{yaxes}{Logical indicating whether to draw y-axis ticks. Defaults to TRUE.}

\item{grid}{Logical indicating whether to add a grid to the plot. Defaults to FALSE.}

\item{lwd}{Line width for the density curves. Defaults to 1.}

\item{lty}{Line type for the density curves. Defaults to 1 (solid).}

\item{...}{Additional graphical parameters to be passed to plot function.}
}
\value{
A plot displaying the density model.
}
\description{
This function scales and normalises KDEs to enable plots pf multiple Kernel Density Estimation (KDE) models on the same frame,
arranging them using y-axis offsets to avoid overlap.
}
\details{
It is intended that a future version of this function will take a list of \code{MCd} objects and plot them with less manual intervention.
}
\examples{
data(BIRE)
# Calculate density models for different site types in Ireland
IR_S<-MCdensity(dl=BIRE[BIRE$Where=='Ireland' & BIRE$ccode=='S',2:3])
IR_P<-MCdensity(dl=BIRE[BIRE$Where=='Ireland' & BIRE$ccode=='P',2:3])
IR_E<-MCdensity(dl=BIRE[BIRE$Where=='Ireland' & BIRE$ccode=='E',2:3])
IR_M<-MCdensity(dl=BIRE[BIRE$Where=='Ireland' & BIRE$ccode=='M',2:3])
IR_F<-MCdensity(dl=BIRE[BIRE$Where=='Ireland' & BIRE$ccode=='F',2:3])
IrD <-MCdensity(dl=BIRE[BIRE$Where=='Ireland',2:3])
# Set up the plot
plot_MCd_offsets(IrD, xlim=c(-6300,1800), ylim=c(0,5.5),xaxt='n')
ax()
# Add the density models for different contexts
plot_MCd_offsets(IR_S, add=T, col=2,lwd=1.5, yoffset=1,yaxt_side=4)
plot_MCd_offsets(IR_P, add=T, col=3,lwd=1.5, yoffset=2-0.1,yaxt_side=2)
plot_MCd_offsets(IR_E, add=T, col=4,lwd=1.5, yoffset=3-0.2,yaxt_side=4)
plot_MCd_offsets(IR_M, add=T, col=5,lwd=1.5, yoffset=4-0.3,yaxt_side=2)
plot_MCd_offsets(IR_F, add=T, col=6,lwd=1.5, yoffset=5-0.4,yaxt_side=4)

}
\seealso{
\code{\link{plot.MCd}}
}
\author{
T. Rowan McLaughlin
}
