% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/revcal.R
\name{revcal}
\alias{revcal}
\title{Reverse Calibrating Function}
\usage{
revcal(N, calcurve = intcal, BC = TRUE)
}
\arguments{
\item{N}{Numeric value representing a year cal. BC (if BC = TRUE) or cal. BP (if BC = FALSE).}

\item{calcurve}{Calibration curve data frame containing columns 'calBP' (calendar years Before Present) and 'age' (radiocarbon years).}

\item{BC}{Logical value indicating whether the input value is in cal. BC (TRUE, default) or cal. BP (FALSE).}
}
\value{
A numeric value representing the reverse-calibrated date in radiocarbon years (if BC = TRUE).
}
\description{
This function performs "reverse calibration" or "uncalibration" t owork out where a calendar year intersections the calibration curve, and returns the radiocarbon year BP closest to that point.
}
\details{
This function searches the calibration curve for the closest match to the input date and returns the corresponding date from the other scale.
}
\examples{
# Reverse calibrate radiocarbon years to calendar years
revcal(1000, intcal)

}
