% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rowcal.R
\name{rowcal}
\alias{rowcal}
\title{Calibrate radiocarbon date with output in rows of two columns}
\usage{
rowcal(date, sigma, calcurve = intcal, BC = TRUE, norm = TRUE)
}
\arguments{
\item{date}{A radiocarbon determination in conventional radiocarbon years BP.}

\item{sigma}{The associated laboratory error.}

\item{calcurve}{The calibration curve object to use for calibration. Defaults to \code{intcal}.}

\item{BC}{Output in calendar years BCE/CE (also known as BC/AD)? Defaults to TRUE.}

\item{norm}{Should be probability masses be normalised? Defaults to TRUE.}
}
\description{
This function calculates the probability masses of a radiocarbon determination given its associated error term and the calibration curve to compare against.
}
\examples{
rowcal(5000,30) # Calibrate 5000±30 BP
}
\references{
McLaughlin, T.R. 2019. On applications of space-time modelling with open-source 14C age calibration. Journal of Archaeological Method and Theory 26, 479–501. https://doi.org/10.1007/s10816-018-9381-3
}
\keyword{calibration}
\keyword{radiocarbon}
