% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rowcalmedian.R
\name{rowcalmedian}
\alias{rowcalmedian}
\title{Find Median of a Radiocarbon Determination}
\usage{
rowcalmedian(date, sigma, ...)
}
\arguments{
\item{date}{Numeric vector of radiocarbon determinations.}

\item{sigma}{Numeric vector of sigma values corresponding to the radiocarbon determinations.}

\item{...}{Additional arguments to be passed to other functions.}
}
\value{
The median radiocarbon determination.
}
\description{
This function calculates the median of a given radiocarbon determination, based on the rowcal function.
}
\details{
This function internally calls the \code{rowcal} function to compute intermediate results.
It then calculates the median radiocarbon determination based on these results.
}
\examples{
date <- c(100, 200, 300)
sigma <- c(0.1, 0.2, 0.3)
rowcalmedian(date, sigma)

}
