% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rowcalmode.R
\name{rowcalmode}
\alias{rowcalmode}
\title{Find Modal Calibrated Date for a radiocarbon determination}
\usage{
rowcalmode(date, sigma, method = "pick", ...)
}
\arguments{
\item{date}{An uncalibrated radiocarbon determination}

\item{sigma}{Sigma values corresponding to the radiocarbon determination}

\item{method}{Method for handling multi-modal probability distributions. Options are 'pick' (default), 'mean', or 'all'.}

\item{...}{Additional arguments to be passed to the \code{rowcal} function.}
}
\value{
The modal calibrated date(s) depending on the chosen method.
}
\description{
This function calculates the modal calibrated date for a given radiocarbon determination, based on the rowcal function.
}
\details{
For multi-modal probability distributions, this function can handle three methods:
\itemize{
\item 'pick': Randomly picks one modal value.
\item 'mean': Calculates the mean of all modal values.
\item 'all': Returns a vector containing all modal values.
}
}
\examples{
# Finding modal calibrated date using default method 'pick'
rowcalmode(date, sigma)

# Finding modal calibrated date using 'mean' method
rowcalmode(date, sigma, method = 'mean')

# Finding all modal calibrated dates
rowcalmode(date, sigma, method = 'all')

}
