% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rowcalsam.R
\name{rowcalsam}
\alias{rowcalsam}
\title{Calendar age samples for a radiocarbon determination}
\usage{
rowcalsam(date, sigma, calcurve = intcal, N = 1, ...)
}
\arguments{
\item{date}{A radiocarbon determination.}

\item{sigma}{Sigma values corresponding to the radiocarbon determination.}

\item{calcurve}{Calibration curve to be used. Default is \code{intcal}.}

\item{N}{Number of age samples to generate. Default is 1.}

\item{...}{Additional arguments to be passed to the \code{rowcal} function.}
}
\value{
A numeric vector containing the age sample(s).
}
\description{
This function calculates one or more calendar age samples for a given radiocarbon determination, based on the \code{rowcal} function.
}
\details{
This function internally calls the \code{rowcal} function to compute a probability distribution for the date.
It then draws one or more age samples based this using inverse transform sampling.
}
\examples{
rowcalsam(1337, 30)

# Generate 5 age samples
rowcalsam(1337, 30, N = 5)

}
