% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rowcalsum.R
\name{rowcalsum}
\alias{rowcalsum}
\title{Calculate a Summed Probability Distribution (SPD) for a Set of Radiocarbon Dates}
\usage{
rowcalsum(DATA = CLIP(), cc = "intcal", sumnorm = TRUE, ...)
}
\arguments{
\item{DATA}{A two- or three-data frame containing radiocarbon date information (date BP, laboratory error, calibration curve). If not provided, it reads data from the clipboard. The optional third column contains the name of the calibration curve object (default is 'intcal').}

\item{cc}{The calibration curve to use. Default is 'intcal'.}

\item{sumnorm}{Logical. If \code{TRUE}, normalizes the summed probability distribution by dividing by the number of dates. Default is \code{TRUE}.}

\item{...}{Additional arguments to be passed to the \code{rowcal} function.}
}
\value{
A two-column matrix. The first column contains the sequence of years, and the second column contains the summed (and optionally normalized) density estimates for each year.
}
\description{
This function is a wrapper for \code{sum.rowyears} generating a SPD without an intermediate step of creating a \code{rowyears} object.
}
\details{
This function calculates the SPD for a set of radiocarbon dates. It generates a time matrix from the input data using the 'timematrix' function, then calculates the summed probability distribution by summing across the time matrix.
If \code{norm} is TRUE, the summed probability distribution is normalized by dividing by the number of dates.
}
\examples{
# Calculate the SPD for a set of radiocarbon dates
dates <- data.frame(dates = c(4840, 4885, 4739, 4826, 4610),
                    sigma = c(45, 50, 27, 24, 31))
spd<-rowcalsum(dates)

# NB this is identical to sum(rowcal(dates)) or rowcal(dates) |> sum()

# Calculate the SPD without normalization of the individual dates
nspd<-rowcalsum(dates, sumnorm=TRUE, norm = FALSE)

Compare the results
plot(spd)
plot(nspd, add=T, type='l', col=1)
legend('topright',fill=c(2,NA),border=c(NA,1),
        legend=c('Normalized','Not normalized'))
[`sum.rowyears`] [`plot.rowyears`]
}
\references{
McLaughlin, T.R. 2019. On applications of space-time modelling with open-source 14C age calibration. Journal of Archaeological Method and Theory 26, 479–501. https://doi.org/10.1007/s10816-018-9381-3
}
\keyword{SPD}
\keyword{calibration}
\keyword{radiocarbon}
