% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rowunif.R
\name{rowunif}
\alias{rowunif}
\title{The uniform distribution for a set of dates}
\usage{
rowunif(x0, x1)
}
\arguments{
\item{x0}{A vector of starting years}

\item{x1}{A vector of ending years}
}
\value{
A list of two-column matrices
}
\description{
This function creates a list of two-column matrices for each unif element matching the output format of \code{rowcal} for multiple dates.
}
\examples{
udates<-rowunif(c(1000, 1090), c(1050, 1100))
MCsam(udates) # random samples
}
\seealso{
\code{\link{rowcal}}
}
\author{
T. Rowan McLaughlin
}
