% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rowyears_operations.R
\name{rowyears_operations}
\alias{rowyears_operations}
\alias{[.rowyears}
\alias{c.rowyears}
\alias{c.rowyear}
\title{Data Operations for 'rowyears' Objects}
\usage{
\method{[}{rowyears}(x, ...)
}
\arguments{
\item{A}{A \code{rowyears} object}
}
\value{
A \code{rowyears} object, represented as a list of two-column matrices. A subset of length 1 is a \code{rowyear} object, a single two-column matrix.
}
\description{
These functions allow subsetting and concatenation for objects of class \code{rowyears}.
}
\details{
Each function operates on objects of class \code{rowyears}, \code{[} for the usual subsetting and \code{c()} for concatenation.

The functions return \code{rowyear} class and includes an updated \code{N} attribute.

\describe{
\item{\code{[.rowyears}}{Subsetting for \code{rowyears} objects.}
\item{\code{c.rowyears}}{Concatenation for \code{rowyears} objects.}
\item{\code{c.rowyear}}{Re-direction to \code{c.rowyears} for \code{rowyear} objects.}

}
}
\examples{
# Example rowyear objects
J <- rowcal(c(4500,4600,4700),c(50,50,50))
Z <- rowcal(4800,34)
plot(J[1:2])
plot(c(J,Z))
}
\seealso{
\code{\link{rowcal}} \code{\link{plot.rowyears}}
}
\author{
T. Rowan McLaughlin
}
