% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SPDboot.R
\name{SPDboot}
\alias{SPDboot}
\title{Calculate Summed Probability Distribution (SPD) Confidence Intervals with Bootstrapping}
\usage{
SPDboot(
  SPD,
  Nboot = 100,
  plot.new = FALSE,
  addlines = FALSE,
  col = rgb(0.8, 0, 0, 0.1),
  ...
)
}
\arguments{
\item{SPD}{A \code{c14sum} object representing the original SPD.}

\item{Nboot}{Number of bootstrap iterations (default is 100).}

\item{plot.new}{Logical indicating whether to create a new plot (default is FALSE).}

\item{addlines}{Logical indicating whether to add lines to an existing plot (default is FALSE).}

\item{col}{Colour for the bootstrapped lines (default is a translucent red).}

\item{...}{Additional graphical parameters passed to the \code{plot} function.}
}
\value{
A matrix representing the bootstrapped SPDs with each column representing a bootstrap iteration.
}
\description{
This function uses bootstrapping to calculate confidence intervals for a given Summed Probability Distribution (SPD).
}
\details{
This function calculates confidence intervals for an SPD using a bootstrapping method (as described by Fernandez-Lopez de Pablo et al 2019). It creates a bootstrapped SPD by randomly selecting calendar years from the original SPD and using them to generate new radiocarbon dates with corresponding standard deviations. This process is repeated for a specified number of bootstrap iterations, resulting in a matrix containing the bootstrapped SPDs.

The function includes options for plotting the bootstrapped SPDs and adding lines to an existing plot.
}
\examples{
# Calculate bootstrapped confidence intervals for an SPD
# Calculate the SPD for a set of radiocarbon dates
dates <- data.frame(BP = c(4840, 4885, 4739, 4826, 4610), sd = c(45, 50, 27, 24, 31))
S<-rowcalsum(dates)
Sboot<-SPDboot(S)
plot(Sboot)

[`rowcalsum`] [`plot.SPDboot`]
}
\references{
Fernández-López de Pablo, J., Gutiérres-Roig, M.,Gómez-Puche, M., Silva, F., McLaughlin, R., and Lozano, S. 2019. Palaeo-demographic modelling supports a population bottleneck during the Pleistocene-Holocene transition in Iberia. Nature Communications 10, 1872. http://doi.org/10.1038/s41467-019-09833-3
}
\keyword{SPD}
\keyword{calibration}
\keyword{radiocarbon}
