% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/suggest_bw.R
\name{suggest_bw}
\alias{suggest_bw}
\title{Suggest KDE bandwidth for radiocarbon determinations using R's Default Methods}
\usage{
suggest_bw(DATA = CLIP(), calcurve = "intcal")
}
\arguments{
\item{DATA}{Data frame containing radiocarbon dates. Columns should be: date, sd, \link{calcurve}.}

\item{calcurve}{A string containing a the name of a calibration curve object.}
}
\value{
A named vector containing suggested bandwidths using various R default methods.
}
\description{
This function suggests bandwidths for kernel density estimation of radiocarbon determinations using R's default methods.
}
\details{
This function calculates suggested bandwidths for kernel density estimation using R's default methods.
It takes as input a data frame containing radiocarbon dates along with their standard deviations and, optionally,
the calibration curve name or object. If not provided, 'intcal' is used as the default calibration curve.
}
\examples{
# Suggest bandwidth using default methods
dates <- data.frame(BP = c(4840, 4885, 4739, 4826, 4610), sd = c(45, 50, 27, 24, 31))
suggest_bw(dates)

}
\references{
Silverman, B. W. (1986). Density Estimation. London: Chapman and Hall.
}
\seealso{
\code{\link{bw.nrd}}
}
