% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sum.rowyears.R
\name{sum.rowyears}
\alias{sum.rowyears}
\title{Sum age-calibrated probability masses}
\usage{
\method{sum}{rowyears}(L, sumnorm = TRUE, na.rm = TRUE)
}
\arguments{
\item{L}{A list of two-column matrices. The first column represents years,
and the second column represents density estimates. Each matrix must have
at least two columns.}

\item{na.rm}{Logical. If \code{TRUE} (default), removes \code{NA} values from the probabilities while computing the sum.}

\item{norm}{Logical. If \code{TRUE} (default), normalizes the sum by dividing by
the number of matrices in the list.}
}
\value{
A two-column matrix. The first column contains the sequence of years,
and the second column contains the summed (and optionally normalized) density
estimates for each year.
}
\description{
Computes the summed probablity for a list of two-column matrices
over a common range of years. Optionally normalizes the results.
}
\details{
This function takes a list of two-column matrices, determines the
overall range of years across all matrices, interpolates the density estimates
for each year, and computes the sum of density estimates for each year. If
normalization is enabled, the sum is divided by the number of matrices in the list.
The probablity masses should thus sum to \code{1} if normalised or \code{length(L)} if not;  if they
don't this indicates problem with the input data having different normalisation
conditions.
}
\examples{
# Example list of two-column matrices
dates <- rowcal(date = c(4840, 4885, 4739, 4826, 4610),
                sigma = c(45, 50, 27, 24, 31))
result <- sum(dates)
plot(result)

}
\author{
T. Rowan McLaughlin
}
