% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timematrix.R
\name{timematrix}
\alias{timematrix}
\title{Generate a Time Matrix from Radiocarbon Date List}
\usage{
timematrix(L)
}
\arguments{
\item{L}{A list of two-column matrices, e.g. of type \code{rowyears} where each matrix represents an age estimate with the first column as the calibrated date and the second column as the probability.}
}
\value{
A for-element list of type \code{timematrix} containing the date list, the calibrated years range, the time matrix, and the specified resolution.
}
\description{
This function generates a time matrix from a list of radiocarbon dates, providing a matrix with calibrated dates for each input date in the list.
}
\details{
This function calculates a time matrix from a list of radiocarbon dates. It first determines the range of calibrated dates based on the input data and user-defined parameters. Then, it generates a 'time grid', where each column represents the posterior probability of a sample's chronology, and each row represents the a year that is the midpoint of calendar sequence with a specified resolution. For each radiocarbon date, it calculates the posteriors using the 'rowcal' function and populates the time matrix accordingly. The column sums are thus the summed probability distribution ('SPD') for the dates.
}
\examples{
# Generate a time matrix from a radiocarbon date list
dates <- rowcal(date = c(4840, 4885, 4739, 4826, 4610),
                sigma = c(45, 50, 27, 24, 31))

test_timematrix<-timematrix(dates)
colSums(test_timematrix,na.rm = TRUE) # each should approximately equal 1
}
\author{
T. Rowan McLaughlin
@seealso
\code{\link{rowcalsum}}
}
