% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/year_densities.R
\name{year_densities}
\alias{year_densities}
\title{Extract Probability Densities for a Given Year}
\usage{
year_densities(Y, MCd)
}
\arguments{
\item{Y}{The year for which the densities are extracted.}

\item{MCd}{A Monte Carlo density-like object (MCd).}
}
\value{
A vector containing the probability densities for the specified year.
}
\description{
This function extracts the probability densities for a given year from an object of class MCd.
}
\examples{
# Extract probability densities for the year 1950
year_densities(1950, denmod)

}
