% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findmedian.R
\name{findmedian}
\alias{findmedian}
\title{Find the median date for each row of a table of dates with different calibration curves.}
\usage{
findmedian(L)
}
\arguments{
\item{L}{A list of matrices containing radiocarbon determinations, calendar dates. Columns are: date, probablitiy.}
}
\value{
A vector containing the median date of each element in the list provided.
}
\description{
This function returns the 'median' date for a list of radiocarbon determinations. The median is calculated by finding the date that corresponds to the 50th percentile of the cumulative probability.
}
\examples{
L<-rowcal(date = c(-3800, 4990, 5300), sigma = c(10, 30, 50),
          cc = c('calcal', 'intcal', 'marine'))
findmedian(L)

}
\seealso{
\code{\link{rowcal}} \code{\link{findmode}} \code{\link{findwm}}
}
\author{
T. Rowan McLaughlin
}
