% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findmode.R
\name{findmode}
\alias{findmode}
\title{Find the mode for Each Row in a Table of Uncalibrated Dates}
\usage{
findmode(L, ...)
}
\arguments{
\item{L}{A two-column list (radiocarbon determination BP, standard deviation) containing uncalibrated radiocarbon dates.}

\item{...}{Additional arguments to be passed to the \code{rowcalmode} function.}
}
\value{
A vector containing the best guess weighted mean date for each uncalibrated radiocarbon date.
}
\description{
This function generates a best guess for a table of uncalibrated radiocarbon dates using the modal value.
}
\details{
This function processes each row of the input list and generates the best guess date using the \code{rowcalmode} function.
}
\examples{
L <- data.frame(BP = c(4100, 4200, 3900), SD = c(20, 34, 50))
findmode(L)

}
