% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findwm.R
\name{findwm}
\alias{findwm}
\title{Find the Weighted Mean for Each Row in a Table of Uncalibrated Dates}
\usage{
findwm(L, ...)
}
\arguments{
\item{L}{A two-column list (radiocarbon determination BP, standard deviation) containing uncalibrated radiocarbon dates.}

\item{...}{Additional arguments to be passed to the \code{rowcalwm} function.}
}
\value{
A vector containing the best guess weighted mean date for each uncalibrated radiocarbon date.
}
\description{
This function generates a best guess for a table of uncalibrated radiocarbon dates.
}
\details{
This function processes each row of the input list and generates the best guess date using the \code{rowcalwm} function.
}
\examples{
L <- data.frame(BP = c(4100, 4200, 3900), SD = c(20, 34, 50))
findwm(L)

}
