% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MCdensity.list.R
\name{MCdensity.list}
\alias{MCdensity.list}
\title{Calculate Gaussian Kernel Density Estimation (KDE) for a list of age-calibrated probability distributions}
\usage{
MCdensity.list(L, N = 100, bw = 30, buffer = 0)
}
\arguments{
\item{L}{List of data frames where each element represents a two-column probability distribution. The first column is time, and the second column is probability.}

\item{N}{Number of Monte Carlo iterations. Default is 100.}

\item{bw}{bw The smoothing bandwidth for density estimation. Default is 30 years. Can also be a character string for automatic selection, ee ‘bw.nrd’.}

\item{buffer}{Buffer to add to the range of time values for density estimation.}
}
\value{
A matrix containing the Gaussian Kernel Density Estimation for each Monte Carlo iteration, allowing summary statistics.
}
\description{
This function calculates the Gaussian Kernel Density Estimation (KDE) for a list of two-column probability distributions, where each element represents a probability distribution (time, probability).
}
\details{
This function performs Gaussian Kernel Density Estimation using a Monte Carlo method for a list of two-column probability distributions.
It converts each element of the input list to a data frame suitable for density estimation, calculates the density estimate for each Monte Carlo iteration, and returns a matrix in the same format as \code{MCdensity} containing the density estimates for further analysis.
}
\examples{
# Calculate Gaussian KDE for a list of two-column probability distributions
 L <- list(rowcal(5310, 35), rowcal(5200, 41), calen(-3900,30),rowcal(5753, 35), rowcal(5500, 41))
denmod <- MCdensity(L)
plot(denmod)

}
\seealso{
\code{\link{MCdensity}} \code{\link{plot.MCd}} \code{\link{density}} \code{\link{bw.nrd}}
}
