% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MCrumned.R
\name{MCrunmed}
\alias{MCrunmed}
\title{Simulate and Compute Running Median for Chronological Data}
\usage{
MCrunmed(
  cal = NULL,
  unif = NULL,
  x = list(),
  y = NULL,
  calcurve = "intcal",
  N = 100,
  y_blur = 0,
  k = 11
)
}
\arguments{
\item{cal}{A matrix of radiocarbon data with 3 or 4 columns:
\itemize{
\item Column 1: Radiocarbon years (14C age).
\item Column 2: Standard deviation of radiocarbon years.
\item Column 3 (optional): Calibration curve names. Defaults to \code{calcurve}.
\item Column 4: Response variable.
}
Default is \code{NULL}.}

\item{unif}{A matrix of uniform chronological data with 3 columns:
\itemize{
\item Column 1: Start of uniform interval (x0).
\item Column 2: End of uniform interval (x1).
\item Column 3: Response variable (y).
}
Default is \code{NULL}.}

\item{x}{A list of calibrated chronological probability data such as the output of \code{rowcal} or \code{rowunif}. Default is an empty list.}

\item{y}{A vector of response variable values associated with \code{x}. Default is \code{NULL}.}

\item{calcurve}{A character string specifying the calibration curve to use when processing
radiocarbon data. Default is "intcal".}

\item{N}{The number of Monte Carlo simulations to perform. Default is 100.}

\item{y_blur}{The standard deviation of Gaussian noise added to \code{y} values. Default is 0 (no noise).}

\item{k}{The width of the running median window (number of points). Default is 11.}
}
\value{
A list of class \code{MCr}, where each element is a two-column matrix containing:
\itemize{
\item Simulated chronological values (column 1).
\item Corresponding running medians of the response variable (column 2).
}
}
\description{
The \code{MCrunmed} function combines input chronological data and associated response
variables, performs Monte Carlo simulations to generate random samples, and computes
a running median for the response variable across the simulated data. There are multiple options for specifing input data, including \code{cal}, \code{unif}, \code{x}, and \code{y}.
}
\details{
The function:
\enumerate{
\item Validates and processes input data (\code{cal} and \code{unif}) to generate chronological samples.
\item Simulates chronological data using Monte Carlo methods.
\item Computes the temporal order of simulated data.
\item Applies a running median to the response variable, optionally adding noise to \code{y}.
}

The three options for input, \code{x} and \code{y}, \code{cal}, and \code{unif} are combined if provided. The resulting data is then simulated
and processed together.

The window size \code{k} should be smaller than the number of points otherwise \code{runmed} will change \code{k} to a lower value and provide a warning.

The returned list can be interpolated to a regular time grid with \code{\link{MCr.as.MCd}}
}
\examples{
# Example data
x0 <- c(-5010, -5110, -5350) # from
x1 <- c(-5000, -5100, -5200) # to
y <- c(0.2, 0.7, 0.9)

# Perform Monte Carlo simulation
result <- MCrunmed(unif=matrix(c(x0, x1, y), ncol=3), k=3, N=100)

# Access the first simulation
head(result[[1]])

# Include some radiocarbon-dated points also
dates<-data.frame(BP=c(6890,6203,5994), error=c(30, 20, 40), response=c(1.1, 0.8, 0.5))
result2 <- MCrunmed(cal=dates, unif=matrix(c(x0, x1, y), ncol=3), k=3, N=100)

}
\seealso{
\code{\link{runmed}} \code{\link{MCr.as.MCd}} \code{\link{rowunif}} \code{\link{rowcal}}
}
\author{
T. Rowan McLaughlin
}
