% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phaser.R
\name{phaser}
\alias{phaser}
\title{Identify chronological phases using hierarchical clustering}
\usage{
phaser(siteids, datelist, h = 30)
}
\arguments{
\item{siteids}{A vector of site identifiers for each date.}

\item{datelist}{A list, typically of type \code{rowyears}, each element being a two-column matrix describing a probability distribution in time.}

\item{h}{Height parameter for dendrogram cutting. Default is 30 years.}
}
\value{
A data frame containing the input data with an additional column indicating the assigned time bin for each date.
}
\description{
This function performs phase binning of radiocarbon determinations (or other chronological information) using hierarchical clustering, assigning each date to a phase based the median value of the calibrated dates.
}
\details{
This function performs phase binning of radiocarbon dates by assigning each date to a phase based on median dates.
It calculates the median date for each site and then uses hierarchical clustering to group the dates into phases.
}
\examples{
\dontrun{
# Perform phase binning with custom height parameter of 100 years

data(BIRE)
# contatenate E and N columns to make a siteid
BIRE$siteid<-paste(BIRE$E,BIRE$N,sep="_")
# Isolate burnt mounds from Ireland
bm_data <- BIRE[BIRE$ccode=="F" & BIRE$Where=='Ireland',]
bm_cal <- rowcal(bm_data$BP, bm_data$SD)
# Store output in a new column
bm_data$phasecode <- phaser(bl_data$siteid, bl_cal, h = 100)
}

}
