% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/suggest_bw.R
\name{suggest_bw}
\alias{suggest_bw}
\title{Suggest KDE bandwidth for radiocarbon determinations using R's Default Methods}
\usage{
suggest_bw(L)
}
\arguments{
\item{L}{A list of radiocarbon determinations or similar, e.g. the output of \code{rowcal} or \code{rowunif} with each element being a two-column matrix with dates and standard deviations.}
}
\value{
A named vector containing suggested bandwidths using various R methods.
}
\description{
This function suggests bandwidths for kernel density estimation of radiocarbon determinations using R's default methods.
}
\details{
This function calculates suggested bandwidths for kernel density estimation using R's built-in alternative algorithms.
It takes as input a data frame containing radiocarbon dates along with their standard deviations and, optionally,
the calibration curve name or object. If not provided, 'intcal' is used as the default calibration curve.
}
\examples{
# Suggest bandwidth using default methods
dates <- rowcal(c(4840, 4885, 4739, 4826, 4610), c(45, 50, 27, 24, 31))
suggest_bw(dates)

# Compare with the `bandwidth.nrd` function in the MASS package:

MASS::bandwidth.nrd(findmedian(dates))

}
\references{
Silverman, B. W. (1986). Density Estimation. London: Chapman and Hall.
}
\seealso{
\code{\link{bw.nrd}} \code{\link{bw.nrd0}} \code{\link{bw.ucv}} \code{\link{bw.bcv}} \code{\link{bw.SJ}} \code{\link{bw.SJ-ste}} \code{\link{bw.SJ-dpi}}
}
\author{
T. Rowan McLaughlin
}
