% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timematrix.R
\name{timematrix}
\alias{timematrix}
\title{Generate a Time Matrix from Radiocarbon Date List}
\usage{
timematrix(
  datelist,
  buffer = 100,
  BC = TRUE,
  res = 10,
  default_curve = "intcal",
  normdates = TRUE,
  quiet = FALSE
)
}
\arguments{
\item{datelist}{A data frame containing radiocarbon date information. It should have at least two columns: the first column representing the radiocarbon dates (BP), the second column representing their corresponding uncertainties. If a third column is present, it should contain the name of the calibration curve onjext (default is 'intcal').}

\item{buffer}{The buffer added to the minimum and maximum calibrated dates to define the range (default is 100).}

\item{BC}{Logical value indicating whether dates are in BC (TRUE) or AD (FALSE) format (default is TRUE).}

\item{res}{The resolution for the time grid (default is 10 years).}

\item{default_curve}{The default calibration curve name to use if not provided in the input data (default is 'intcal').}

\item{quiet}{Logical value indicating whether to suppress progress bars (default is FALSE). The matrix will take a few moments to assemble for many 100s of thousands of dates.}
}
\value{
A for-element list of type \code{timematrix} containing the radiocarbon date list, the calibrated years range, the time matrix, and the specified resolution.
}
\description{
This function generates a time matrix from a list of radiocarbon dates, providing a matrix with calibrated dates for each input date in the list.
}
\details{
This function calculates a time matrix from a list of radiocarbon dates. It first determines the range of calibrated dates based on the input data and user-defined parameters. Then, it generates a 'time grid', where each column represents the posterior probability of a sample's chronology, and each row represents the a year that is the midpoint of calendar sequence with a specified resolution. For each radiocarbon date, it calculates the posteriors using the 'rowcal' function and populates the time matrix accordingly. The column sums are thus the summed probability distribution ('SPD') for the dates.
}
\examples{
# Generate a time matrix from a radiocarbon date list
dates <- data.frame(BP = c(4840, 4885, 4739, 4826, 4610), sd = c(45, 50, 27, 24, 31))

test_timematrix<-timematrix(dates)

 @seealso
[`rowcalsum`] 
}
